# controllers/Hospitalmng_db/generated/RegistrationControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE RegistrationControllerGenerated.js PLEASE EDIT ../RegistrationController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import RegistrationModel from "../../../models/Hospitalmng_db/RegistrationModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import RegistrationController from "../RegistrationController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/registration`;
    router.post(baseUrl + "", authorize([]), RegistrationController.create);
    router.delete(baseUrl + "/:id", authorize([]), RegistrationController.delete);
    router.get(baseUrl + "/findBy_Doctor/:key", authorize([]), RegistrationController.findBy_Doctor);
    router.get(baseUrl + "/findBy_patients/:key", authorize([]), RegistrationController.findBy_patients);
    router.get(baseUrl + "/findBy_user/:key", authorize([]), RegistrationController.findBy_user);
    router.get(baseUrl + "/:id", authorize([]), RegistrationController.get);
    router.get(baseUrl + "", authorize([]), RegistrationController.list);
    router.post(baseUrl + "/:id", authorize([]), RegistrationController.update);
  },


  // CRUD METHODS


  /**
  * RegistrationModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await RegistrationModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RegistrationModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await RegistrationModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RegistrationModel.findBy_Doctor
  *   @description CRUD ACTION findBy_Doctor
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_Doctor: async (req, res) => {
    try {
      const result = await RegistrationModel.findBy_Doctor(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RegistrationModel.findBy_patients
  *   @description CRUD ACTION findBy_patients
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_patients: async (req, res) => {
    try {
      const result = await RegistrationModel.findBy_patients(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RegistrationModel.findBy_user
  *   @description CRUD ACTION findBy_user
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_user: async (req, res) => {
    try {
      const result = await RegistrationModel.findBy_user(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RegistrationModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await RegistrationModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * RegistrationModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await RegistrationModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * RegistrationModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await RegistrationModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;